+++ src/Test/Builder/AbstractStep.php
@@ -17,11 +17,4 @@ abstract class AbstractStep implements Rule
     {
         return $this->rule;
     }
-
-    public function nonIgnorable(): static
-    {
-        $this->rule->params['nonIgnorable'] = true;
-
-        return $this;
-    }
 }