+++ src/Test/Builder/BuildStep.php
@@ -2,4 +2,12 @@
 
 namespace PHPat\Test\Builder;
 
-class BuildStep extends AbstractStep {}
+class BuildStep extends AbstractStep
+{
+    public function nonIgnorable(): static
+    {
+        $this->rule->nonIgnorable = true;
+
+        return $this;
+    }
+}