+++ docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+* Update stubs for built-in classes (Sept 2026)
+* 
+* Add `nonIgnorable()` to prevent rule errors from being ignored
+
 ## 0.12.4
 * Split assertions from constraints (`should`, `shouldNot`, `canOnly`) and update rule builder steps
 * Deprecate all previous `should*`, `shouldNot*`, and `canOnly*` assertions